Supersede Older Guidelines
The Board of Directors approved 2020.06 on June 11, 2020.  The
Foundation's OpenStack Powered program currently allows vendors to
use either of the newest two guidelines when testing [1].  The
two most recent Guidelines are now 2020.06 and 2019.11, so we need
to mark older Guideliens as superseded.  This was missed in the last
round of approvals, so this patch corrects their status.

[1] https://www.openstack.org/brand/interop/

Change-Id: I415ddebe37873847c6a272d9249dfa048428d59c